requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
anthropic>=0.40.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0
